requirements.txt
streamlit
python-doctr[torch]
opencv-python
Pillow
numpy<2.0
